refactor: 폰트 상수폴더로 이동
- Txt 컴포넌트 안에 있던 폰트를 다른 컴포넌트에서도 사용하는데 import문 자체가
Txt에서 가져오는 것이 이질적으로 느껴져 상수 폴더로 옮기고 Txt 컴포넌트에서도 import해서
사용하는 방식으로 변경
