+++ fe/src/components/common/DetailTextInput.tsx
@@ -1,9 +1,10 @@
 import { useState } from "react";
 // import { color } from "../../constants/colors";
-import { fonts } from "../util/Txt";
+// import { fonts } from "../util/Txt";
 import { Icon } from "./Icon";
 import { ColorScheme } from "../../contexts/ThemeContext";
 import { useTheme } from "@emotion/react";
+import { fonts } from "../../constants/fonts";
 
 type Props = {
   mode: string;
