+++ fe/src/components/common/TabButton.tsx
@@ -1,8 +1,9 @@
 // import { color } from "../../constants/colors";
 import { useTheme } from "@emotion/react";
 import { ColorScheme } from "../../contexts/ThemeContext";
-import { fonts } from "../util/Txt";
+// import { fonts } from "../util/Txt";
 import { Button } from "./Button";
+import { fonts } from "../../constants/fonts";
 
 export function TabButton({
   isLeftSelected,
