+++ fe/src/constants/fonts.ts
@@ -1,0 +1,78 @@
+const fontFamily = `-apple-system, BlinkMacSystemFont, "Apple SD Gothic Neo",
+"Pretendard Variable", Pretendard, Roboto, "Noto Sans KR", "Segoe UI",
+"Malgun Gothic", "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol",
+sans-serif`;
+
+export const fonts = {
+  bold32: {
+    fontSize: "32px",
+    fontWeight: 700,
+    // lineHeight: "48px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  bold24: {
+    fontSize: "24px",
+    fontWeight: 700,
+    // lineHeight: "36px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  bold20: {
+    fontSize: "20px",
+    fontWeight: 700,
+    // lineHeight: "32px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  bold16: {
+    fontSize: "16px",
+    fontWeight: 700,
+    // lineHeight: "24px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  bold12: {
+    fontSize: "12px",
+    fontWeight: 700,
+    // lineHeight: "16px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  medium32: {
+    fontSize: "32px",
+    fontWeight: 500,
+    // lineHeight: "48px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  medium24: {
+    fontSize: "24px",
+    fontWeight: 500,
+    // lineHeight: "36px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  medium20: {
+    fontSize: "20px",
+    fontWeight: 500,
+    // lineHeight: "32px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+  medium16: {
+    fontSize: "16px",
+    fontWeight: 500,
+    // lineHeight: "24px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+    textAlign: "left" as const,
+  },
+  medium12: {
+    fontSize: "12px",
+    fontWeight: 500,
+    // lineHeight: "16px",
+    letterSpacing: 0,
+    fontFamily: fontFamily,
+  },
+};
